This function allows you to calculate the people enrolled in school

This function allows you to calculate the people enrolled in school

enrolled_school(
  data = ech::toy_ech_2018,
  e193 = "e193",
  e197 = "e197",
  e201 = "e201",
  e212 = "e212",
  e215 = "e215",
  e218 = "e218",
  e221 = "e221",
  e224 = "e224"
)

Arguments

data
data.frame with necessary variables Defaults to ech.

e193
Variable name of attendance school

e197
Variable name of attendance primary

e201
Variable name of attendance secondary

e212
Variable name of attendance technical school (non-university)

e215
Variable name of attendance magisterio

e218
Variable name of attendance university

e221
Variable name of attendance tertiary

e224
Variable name of attendance postgrade

Value

data.frame

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also

Examples

# \donttest{
toy_ech_2018 <- enrolled_school(data = ech::toy_ech_2018)

#> A variable has been created: 
#>  	  school_enrollment (matriculacion escolar)

# }