get_estimation_mean

This function allows you to estimate mean variable at universe level.

get_estimation_mean(
  data = ech::toy_ech_2018,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  name = "estimacion"
)

Arguments

data
data frame with ECH microdata

variable
data frame column to estimate

by.x
data frame column

by.y
data frame column

domain
subpopulation reference setted as logical

level
is household ("h") or individual ("i").

name
name for the estimation new column

Value

table

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

Examples

# \donttest{
get_estimation_mean(variable = "pobre06", by.x = "dpto", level = "h")

#> Warning: The `add` argument of `group_by()` is deprecated as of dplyr 1.0.0.
#> Please use the `.add` argument instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 19 x 3
#>                   dpto colname colname_se
#>              <dbl+lbl>   <dbl>      <dbl>
#>  1  1 [Montevideo]      0.198      0.0272
#>  2  2 [Artigas]         0.143      0.0663
#>  3  3 [Canelones]       0.0457     0.0182
#>  4  4 [Cerro Largo]     0          0     
#>  5  5 [Colonia]         0          0     
#>  6  6 [Durazno]         0          0     
#>  7  7 [Flores]          0          0     
#>  8  8 [Florida]         0.0938     0.0519
#>  9  9 [Lavalleja]       0.0455     0.0445
#> 10 10 [Maldonado]       0.0167     0.0167
#> 11 11 [Paysandú]        0.0645     0.0613
#> 12 12 [Río Negro]       0.105      0.0702
#> 13 13 [Rivera]          0.179      0.0619
#> 14 14 [Rocha]           0.0952     0.0641
#> 15 15 [Salto]           0.0816     0.0473
#> 16 16 [San José]        0.0233     0.0230
#> 17 17 [Soriano]         0          0     
#> 18 18 [Tacuarembó]      0.0370     0.0364
#> 19 19 [Treinta y Tres]  0          0     

# }