This function allows you to estimate median variable at universe level.

This function allows you to estimate median variable at universe level.

get_estimation_median(
  data = ech::toy_ech_2018,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  ids = NULL,
  numero = "numero",
  estrato = NULL,
  pesoano = "pesoano",
  name = "estimacion"
)

Arguments

data
data frame with ECH microdata

variable
data frame column to estimate

by.x
data frame column

by.y
data frame column

domain
subpopulation reference setted as character expresion of logical evaluation

level
is household ("h") or individual ("i").

ids
ids

numero
household id

estrato
strata

pesoano
weights

name
name for the estimation new column

Value

table

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also

Examples

# \donttest{
get_estimation_median(data = ech::toy_ech_2018, variable = "ht11", by.x = "dpto", level = "h")

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> Warning: Only diagonal of vcov() available

#> These confidence intervals are only an approximation of the correct confidence intervals 
#>   that arise from fully defining the sample design

#> # A tibble: 19 x 4
#>    dpto           estimacion estimacion_low estimacion_upp
#>    <fct>               <dbl>          <dbl>          <dbl>
#>  1 Montevideo         75061.         66152.         84688.
#>  2 Artigas            53237.         24011.         83625.
#>  3 Canelones          63574.         46839.         73202.
#>  4 Cerro Largo        38440.         20362.         58701.
#>  5 Colonia            58404.         30456.         87953.
#>  6 Durazno            32050          20080          39968.
#>  7 Flores             67956          67956         110105 
#>  8 Florida            57815.         38537.         73721.
#>  9 Lavalleja          41718.           177.         99331.
#> 10 Maldonado          56626          35256.         83030.
#> 11 Paysandú           42049.         29696.         71206.
#> 12 Río Negro          48668.         19000         138824 
#> 13 Rivera             53175.         33945.         97111.
#> 14 Rocha              44288.         29872.         73749.
#> 15 Salto              50109.         40945.         65198.
#> 16 San José           54366.         41438.         58834.
#> 17 Soriano            45274.         27778.         87152.
#> 18 Tacuarembó         40920.         26784.         57952.
#> 19 Treinta y Tres     31532.         14050          80450 

# }