This function allows you to estimate de Income Quintile Share Ratio

This function allows you to estimate de Income Quintile Share Ratio

get_estimation_qsr(
  data = ech::toy_ech_2018,
  variable = "y_pc_d_r",
  by = NULL,
  ids = NULL,
  estrato = NULL,
  pesoano = "pesoano"
)

Arguments

data

data.frame

variable

Variable name of total income per hour

by

data frame column

ids

Variable name of cluster

estrato

Variable name of strata

pesoano

Variable name of weights

Value

table

See also

Examples

# \donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018, index = "IPC", level = "R",
                                       base_month = "01", base_year = "2005")
#> Variables have been created: 
#>  	 y_pc (income per capita current prices / ingreso per capita a precios corrientes)
#>                 y_pc_d_r (income per capita deflated / ingreso per capita deflactado);
#>                 rv_d_r (rental value deflated / valor locativo deflactado);
#>                 y_wrv_d_r (income without rental value deflated / ingreso sin valor locativo deflactado) &
#>                 y_wrv_pc_d_r (income without rental value per capita deflated / ingreso sin valor locativo per capita deflactado)
get_estimation_qsr(data = toy_ech_2018, variable = "y_pc_d_r", pesoano = "pesoano")
#>      value
#> 1 6.461858
# }