This function allows you to calculate the housing status

This function allows you to calculate the housing status

housing_deprivation(
  data = ech::toy_ech_2018,
  n = 1,
  ht19 = "ht19",
  d9 = "d9",
  d10 = "d10",
  d11 = "d11",
  d12 = "d12",
  d13 = "d13",
  d16 = "d16",
  d18 = "d18",
  d19 = "d19",
  c2 = "c2",
  c3 = "c3",
  c4 = "c4",
  quintil = "quintil",
  region_4 = "region_4"
)

Arguments

data

data.frame

n

number of deprivations to consider. Default 1

ht19

Variable name of number of individuals in the household

d9

Variable name of number of rooms

d10

Variable name of number of rooms to sleep

d11

Variable name of principal source of potable water

d12

Variable name of water supply network / water access

d13

Variable name of sanitary facilities

d16

Variable name of sewerage facilities

d18

Variable name of energy source for lighting

d19

Variable name of cooking space

c2

Variable name of predominant material on external walls

c3

Variable name of predominant roofing material

c4

Variable name of predominant flooring material

quintil

Variable name of income quintil

region_4

Variable name of region

Value

data.frame

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also

Examples

# \donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018)
#> Variables have been created: 
#>  	 y_pc (income per capita current prices / ingreso per capita a precios corrientes);
#>     y_pc_d  (income per capita deflated / ingreso per capita deflactado);
#>          rv_d (rental value deflated / valor locativo deflactado);
#>          y_wrv_d (income without rental value deflated / ingreso sin valor locativo deflactado) &
#>          y_wrv_pc_d (income without rental value per capita deflated / ingreso sin valor locativo per capita deflactado)
toy_ech_2018 <- income_quantiles(data = toy_ech_2018)
#> A variable has been created: 
#>  	 quintil (quintil de ingresos)
toy_ech_2018 <- housing_deprivation(data = toy_ech_2018)
#> Variables have been created: 
#>  	 overcrowding (Carencia: Hacinamiento);
#>          bathroom (Carencia: Banio (Minimo un banio));
#>          rooms (Carencia: Ambientes adecuados(cocina, comedor, estar diario));
#>          roof_materials (Carencia: Techo adecuado);
#>          wall_materials (Carencia: Paredes adecuadas);
#>          floor_materials (Carencia: Pisos adecuados);
#>          water (Carencia: Agua);
#>          running_water (Carencia: Red general para el agua o pozo);
#>          sewerage (Carencia: Desague);
#>          electricity (Carencia: Red electrica);
#>          housing_deprivation_q (Suma de carencias habitacionales) &
#>          housing_deprivation (Carencia habitacional en relacion al parametro n)
# }