This function allows you to calculate the labor income per capita

This function allows you to calculate the labor income per capita

labor_income_per_capita(
  data = ech::toy_ech_2018,
  numero = "numero",
  pobpcoac = "pobpcoac",
  g126_1 = "g126_1",
  g126_2 = "g126_2",
  g126_3 = "g126_3",
  g126_4 = "g126_4",
  g126_5 = "g126_5",
  g126_6 = "g126_6",
  g126_7 = "g126_7",
  g126_8 = "g126_8",
  g127_3 = "g127_3",
  g128_1 = "g128_1",
  g129_2 = "g129_2",
  g130_1 = "g130_1",
  g131_1 = "g131_1",
  g133_1 = "g133_1",
  g133_2 = "g133_2",
  g134_1 = "g134_1",
  g134_2 = "g134_2",
  g134_3 = "g134_3",
  g134_4 = "g134_4",
  g134_5 = "g134_5",
  g134_6 = "g134_6",
  g134_7 = "g134_7",
  g134_8 = "g134_8",
  g135_3 = "g135_3",
  g136_1 = "g136_1",
  g137_2 = "g137_2",
  g138_1 = "g138_1",
  g139_1 = "g139_1",
  g141_1 = "g141_1",
  g141_2 = "g141_2",
  g142 = "g142",
  g144_1 = "g144_1",
  g144_2_1 = "g144_2_1",
  g144_2_3 = "g144_2_3",
  g144_2_4 = "g144_2_4",
  g144_2_5 = "g144_2_5"
)

Arguments

data
data frame

numero
Variable name of household id

pobpcoac
Variable name of definition of population by activity status

g126_1
Variable name of net salary

g126_2
Variable name of commissions, incentives, overtime payment, fringe benefits

g126_3
Variable name of non-surrendering expenses

g126_4
Variable name of tips

g126_5
Variable name of annual complementary salary

g126_6
Variable name of vacation pay

g126_7
Variable name of delayed payments

g126_8
Variable name of transportation tickets

g127_3
Variable name of received food or drink

g128_1
Variable name of received food tickets

g129_2
Variable name of received housing or accommodation

g130_1
Variable name of another type of compensation

g131_1
Variable name of received another type of supplement paid by the employer

g133_1
Variable name of the right to cultivate goods for own-consumption

g133_2
Variable name of the right to cultivate goods for own-consumption (amount received from the sale)

g134_1
Variable name of net salary

g134_2
Variable name of commissions, incentives, overtime payment, fringe benefits

g134_3
Variable name of non-surrendering expenses

g134_4
Variable name of tips

g134_5
Variable name of annual complementary salary

g134_6
Variable name of vacation pay

g134_7
Variable name of delayed payments

g134_8
Variable name of transportation tickets

g135_3
Variable name of received food or drink

g136_1
Variable name of received food tickets

g137_2
Variable name of received housing or accommodation

g138_1
Variable name of another type of compensation

g139_1
Variable name of received another type of supplement paid by the employer

g141_1
Variable name of the right to cultivate goods for own-consumption

g141_2
Variable name of the right to cultivate goods for own-consumption (amount received from the sale)

g142
Variable name of withdrawals for business household expenses you have or had

g144_1
Variable name of collected products for own consumption (non-agricultural worker)

g144_2_1
Variable name of collected products for own consumption (non-agricultural worker)

g144_2_3
Variable name of collected products for own consumption (non-agricultural worker)

g144_2_4
Variable name of collected products for own consumption (non-agricultural worker)

g144_2_5
Variable name of collected products for own consumption (non-agricultural worker)

Value

data.frame

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also

Examples

# \donttest{
toy_ech_2018 <- labor_income_per_capita(data = ech::toy_ech_2018)

#> Variables have been created: 
#>  	 labor_income (Ingresos laborales) &
#>             labor_income_h (Ingresos laborales del hogar) &
#>             labor_income_h_percapita (Ingresos laborales per capita)

# }