This function allows you to calculate the highest level of education achieved

This function allows you to calculate the highest level of education achieved

level_education(
  data = ech::toy_ech_2018,
  e51_2 = "e51_2",
  e51_3 = "e51_3",
  e51_4 = "e51_4",
  e51_5 = "e51_5",
  e51_6 = "e51_6",
  e51_7 = "e51_7",
  e51_7_1 = "e51_7_1",
  e51_8 = "e51_8",
  e51_9 = "e51_9",
  e51_10 = "e51_10",
  e51_11 = "e51_11",
  e193 = "e193",
  e49 = "e49"
)

Arguments

data

data.frame

e51_2

Variable name of years passed in primary

e51_3

Variable name of years passed in special primary

e51_4

Variable name of years passed in lower secondary

e51_5

Variable name of years passed in upper secondary

e51_6

Variable name of years passed in technical upper secondary

e51_7

Variable name of years passed in technical school

e51_7_1

Variable name of technical school requirements

e51_8

Variable name of years passed in magisterio/profesorado

e51_9

Variable name of years passed in university or similar

e51_10

Variable name of years passed in tertiary (non-university)

e51_11

Variable name of years passed in postgrade

e193

Variable name of attendance school

e49

Variable name of attendance school ever

Value

data.frame

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also

Examples

# \donttest{
toy_ech_2018 <- level_education(data = ech::toy_ech_2018)
#> A variable has been created: 
#>  	 highest level_education achieved (maximo nivel educativo alcanzado)
# }