poverty

poverty

poverty(
  data = ech::toy_ech_2018,
  scale = 0.8,
  region_4 = "region_4",
  dpto = "dpto",
  ht11 = "ht11",
  ht19 = "ht19",
  numero = "numero"
)

Arguments

data
data.frame

scale
equivalency scale

region_4
Variable name of region. Default: region_4

dpto
Variable name of departamento. Default: dpto

ht11
Variable name of income. Default: ht11

ht19
Variable name of number of individuals in the household. Default: ht19

numero
household id

Value

data.frame

Details

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

Examples

toy_ech_2018 <- poverty(data = ech::toy_ech_2018)

#> Variables have been created: 
#>  	 poor (pobre) &
#>               indigent (indigente)