This function allows you to read ECH from a local folder

This function allows you to read ECH from a local folder

read_microdata(path = NULL)

Arguments

path
Folder where are the files or be download

Value

an object called df

Details

Disclaimer: El script no es un producto oficial de INE.

Disclaimer: This script is not an official INE product. Aviso: El script no es un producto oficial de INE.

See also